refactor(extensions): simplify ApplicationOptions and EnvironmentOptions
Removed default value assignments in ApplicationOptions and updated
EnvironmentOptions to use null-forgiving operator for string properties.
This enhances clarity and ensures that the properties are set directly
from the provided options.
