+++ libs/xSdk.Core/src/Extensions/Options/ApplicationOptions.cs
@@ -37,8 +37,7 @@ public sealed class ApplicationOptions
 
     [Variable(
         name: Definitions.AppName.Name,
-        helpText: Definitions.AppName.HelpText,
-        defaultValue: Definitions.AppName.DefaultValue,
+        helpText: Definitions.AppName.HelpText,        
         resourceNames: ["app.name"],
         hidden: true
     )]
@@ -54,8 +53,7 @@ public sealed class ApplicationOptions
 
     [Variable(
         name: Definitions.AppCompany.Name,
-        helpText: Definitions.AppCompany.HelpText,
-        defaultValue: Definitions.AppCompany.DefaultValue,
+        helpText: Definitions.AppCompany.HelpText,        
         resourceNames: ["app.company"],
         hidden: true
     )]
@@ -66,8 +64,7 @@ public sealed class ApplicationOptions
 
     [Variable(
         name: Definitions.AppPrefix.Name,
-        helpText: Definitions.AppPrefix.HelpText,
-        defaultValue: Definitions.AppPrefix.DefaultValue,
+        helpText: Definitions.AppPrefix.HelpText,        
         resourceNames: ["app.prefix"],
         hidden: true
     )]
@@ -81,7 +78,6 @@ public sealed class ApplicationOptions
         {
             public const string Name = nameof(AppName);
             public const string HelpText = "Short name of the application";
-            public const string DefaultValue = "xsdk";
         }
 
         public static class AppDescription
@@ -94,14 +90,12 @@ public sealed class ApplicationOptions
         {
             public const string Name = nameof(AppCompany);
             public const string HelpText = "Company name of the application";
-            public const string DefaultValue = "xcom";
         }
 
         public static class AppPrefix
         {
             public const string Name = nameof(AppPrefix);
             public const string HelpText = "Prefix for the application";
-            public const string DefaultValue = "XSDK";
         }
 
         public static class AppVersion
