+++ libs/xSdk/src/Extensions/Options/OptionsBase.cs
@@ -1,8 +1,8 @@
 using xSdk.Extensions.Variable;
 
-namespace xSdk.Extensions.Plugin;
+namespace xSdk.Extensions.Options;
 
-public abstract class PluginOptionsBase : VariableSetup
+public abstract class OptionsBase : VariableSetup
 {
     public override string ToString() => base.ToString() ?? string.Empty;
 }
