+++ libs/xSdk/src/Extensions/Options/ServiceCollectionExtensions.cs
@@ -44,8 +44,7 @@ public static class ServiceCollectionExtensions
             services
                 .AddOptions<TOptions>(name)
                 .Configure<IVariableService, IValidator<TOptions>>((options, variableService, validator) =>
-                {
-                    variableService.ParseForVariables(options);
+                {                    
                     if (options is VariableSetup variableSetup)
                     {
                         variableSetup.Initialize(variableService);
